python3/json_web_token/pyproject.toml
[project]
name = "json-web-token"
version = "0.1.0"
description = "JWS signing and verification examples with certificates"
authors = [
    { name = "kmilo", email = "kmilo.denis.glez@yandex.com" }
]
dependencies = [
    "certifi>=2022.12.7",
    "cffi>=1.15.1",
    "charset-normalizer>=3.0.1",
    "cryptography>=39.0.0",
    "idna>=3.4",
    "pycparser>=2.21",
    "pydantic>=1.10.5",
    "PyJWT>=2.6.0",
    "requests>=2.28.2",
    "typing_extensions>=4.5.0",
    "urllib3>=1.26.14",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.3.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/json_web_token"]
